feat(bento): Add favorite contacts grid
Introduces a new "Favorites" grid component to the main Bento screen.

This new grid displays a list of favorite contacts as circular avatars. It also includes an "Add" button to allow users to add new contacts.

The component is added to the main screen layout, replacing a placeholder grid item.