feat: Override close button behavior in Send fragment
This change introduces the ability to override the default "close" button action in the `FragmentSend` screen.

Previously, the close button was hardcoded to trigger `onBackPressed()`. Now, a new `overrideOnCloseClicked` function allows parent components, like the `BentoMainScreen`, to provide a custom callback. This is used to correctly dismiss the bottom sheet when the send fragment is presented within it.